Refactor Android Emulator Snapshot CI Workflow
- Updated e2e-android-emulator-snapshot-ci.ts to streamline the check and prime processes, improving environment variable management and error handling.
- Enhanced the CI workflow configurations in ci.yml and prime-appium-e2e-android-emulator.yml to clarify boot modes and snapshot handling.
- Introduced AndroidGoldenSnapshot.ts to encapsulate snapshot-related logic, promoting code reuse and maintainability.
- Refactored EmulatorHelpers.ts to utilize the new AndroidGoldenSnapshot module, removing redundant code and improving clarity.
- Adjusted tests in EmulatorHelpers.test.ts to align with the new structure and ensure accurate snapshot validation.
