+++ .github/workflows/ci.yml
@@ -1727,7 +1727,7 @@ jobs:
       test-timeout-minutes: 25
       build_type: 'main'
       metamask_environment: 'e2e'
-      # Runner-local AVD only — skip golden AVD artifact download.
+      # Skip golden AVD download; boot-mode cold is redundant with that but keeps intent explicit.
       cache-android-golden-snapshot: false
       android-emulator-boot-mode: cold
       # Mirrors build-android-apps' resolution chain so this consumer always matches
